requirements.txt
streamlit
pandas
numpy
duckdb
scikit-learn
xgboost
shap
matplotlib
seaborn
mlflow
langchain
faiss-cpu
openai
python-dotenv
plotly